backend/tests/test_jwt_handler.py
"""
Tests for JWT token generation, decoding, and password hashing.
"""

import pytest
from datetime import timedelta
from unittest.mock import patch
from jose import jwt as jose_jwt

from auth.jwt_handler import (
    create_access_token,
    decode_access_token,
    verify_password,
    get_password_hash,
    pwd_context,
)
from exceptions import AuthenticationException


# ── Password hashing ─────────────────────────────────────────────────────
# NOTE: passlib has a known incompatibility with bcrypt>=4.1 on Python 3.14
# (detect_wrap_bug sends a 255-byte secret that newer bcrypt rejects).
# We test through a mock to validate the application's usage of CryptContext
# while avoiding the upstream bug.


def _bcrypt_available() -> bool:
    """Check if passlib bcrypt backend works in this environment."""
    try:
        get_password_hash("probe")
        return True
    except (ValueError, Exception):
        return False


_skip_bcrypt = pytest.mark.skipif(
    not _bcrypt_available(),
    reason="passlib bcrypt backend broken on this Python/bcrypt version",
)


class TestPasswordHashing:
    @_skip_bcrypt
    def test_hash_and_verify(self):
        raw = "my_secure_password"
        hashed = get_password_hash(raw)
        assert hashed != raw
        assert verify_password(raw, hashed)

    @_skip_bcrypt
    def test_wrong_password_fails(self):
        hashed = get_password_hash("correct")
        assert not verify_password("wrong", hashed)

    @_skip_bcrypt
    def test_hashes_are_unique(self):
        h1 = get_password_hash("same")
        h2 = get_password_hash("same")
        # bcrypt salts should make them different
        assert h1 != h2

    def test_hash_calls_pwd_context(self):
        """Verify get_password_hash delegates to pwd_context.hash (works without bcrypt)."""
        with patch.object(pwd_context, "hash", return_value="$2b$12$fake") as mock_hash:
            result = get_password_hash("secret")
            mock_hash.assert_called_once_with("secret")
            assert result == "$2b$12$fake"

    def test_verify_calls_pwd_context(self):
        """Verify verify_password delegates to pwd_context.verify (works without bcrypt)."""
        with patch.object(pwd_context, "verify", return_value=True) as mock_verify:
            result = verify_password("plain", "$2b$12$hash")
            mock_verify.assert_called_once_with("plain", "$2b$12$hash")
            assert result is True


# ── Token creation / decoding ────────────────────────────────────────────


class TestTokenCreation:
    def test_roundtrip(self, jwt_secret, jwt_algorithm):
        payload = {"sub": "user-1", "company_id": "comp-1"}
        token = create_access_token(data=payload)
        decoded = decode_access_token(token)

        assert decoded["sub"] == "user-1"
        assert decoded["company_id"] == "comp-1"
        assert "exp" in decoded

    def test_custom_expiry(self, jwt_secret, jwt_algorithm):
        token = create_access_token(
            data={"sub": "u"}, expires_delta=timedelta(minutes=5)
        )
        decoded = decode_access_token(token)
        assert decoded["sub"] == "u"

    def test_expired_token_raises(self, jwt_secret, jwt_algorithm):
        token = create_access_token(
            data={"sub": "u"}, expires_delta=timedelta(seconds=-1)
        )
        with pytest.raises(AuthenticationException):
            decode_access_token(token)

    def test_tampered_token_raises(self, jwt_secret, jwt_algorithm):
        token = create_access_token(data={"sub": "u"})
        tampered = token[:-4] + "XXXX"
        with pytest.raises(AuthenticationException):
            decode_access_token(tampered)

    def test_garbage_token_raises(self):
        with pytest.raises(AuthenticationException):
            decode_access_token("not.a.jwt")
