backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.1
pydantic>=2.7.4
pydantic-settings>=2.3.4
email-validator>=2.1.0
langchain>=0.2.6
langchain-openai>=0.1.13
langchain-google-genai>=1.0.6
mlflow>=2.14.1
boto3>=1.34.135
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
slowapi>=0.1.9
python-dotenv>=1.0.1
pytest>=8.0.0
pytest-asyncio>=0.23.0
httpx>=0.27.0
duckduckgo-search>=7.0.0
